Cleanup database initialization

+++ src/Tests/conftest.py
@@ -48,8 +48,10 @@ with open(os.path.join(os.path.dirname(__file__), "data_default.sql"), "rb") as 
 
 
 @pytest.fixture
-def app():
-    db_fd, db_path = tempfile.mkstemp()
+def app(db_path=None):
+    db_is_new = db_path is None
+    if db_is_new:
+        db_fd, db_path = tempfile.mkstemp()
 
     app = LifeLogServer.create_app(
         config_file=DEFAULT_CONFIG_FILE,
@@ -57,13 +59,13 @@ def app():
     )
 
     with app.app_context():
-        LifeLogServer.database.init_db()
-        LifeLogServer.database.get_db(new_connection=True).executescript(_data_sql)
+        LifeLogServer.database.get_db().executescript(_data_sql)
 
     yield app
 
-    os.close(db_fd)
-    os.unlink(db_path)
+    if db_is_new:
+        os.close(db_fd)
+        os.unlink(db_path)
 
 
 @pytest.fixture
